requirements.txt
crewai
openai
langchain
langchain-openai
yfinance
ta
pandas
numpy
python-dotenv
aiohttp
streamlit
plotly